src/backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Optional: Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Optional: Utilities
requests==2.31.0
Pillow==10.1.0

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
